pyproject.toml
[project]
name = "rag-knowledge"
version = "0.1.0"
description = "RAG Agent with LangChain and Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-google-genai>=1.0.0",
    "langchain-text-splitters>=0.0.1",
    "langchain-core>=0.1.0",
    "beautifulsoup4>=4.12.0",
    "chromadb>=0.4.22",
    "python-dotenv>=1.0.0",
    "lxml>=5.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pypdf>=4.0.0",
    "docx2txt>=0.8",
    "duckduckgo-search>=8.1.1",
    "ddgs>=9.9.1",
]

[tool.uv]
dev-dependencies = [
    "ipython>=8.12.0",
]
